started adding functions to start, send and request session and prompt
Started dealing with request and prompt handling within send_LLM(); also added additional send_Http_request(); in order to start session with open source LLM and request answer based on user input. (Pretty sure this is able to be simplified into only a send_LLM(); will come back and brainstorm that. Kind of want it to use a seperate request function for orginazation and readability with this project though!